Bugfix - When get a pokemon for the first time it arrives without stages

+++ src/main/java/com/pokedexplus/poke_dex_plus_backend/repository/EvolutionChainRepository.java
@@ -3,5 +3,5 @@ package com.pokedexplus.poke_dex_plus_backend.repository;
 import com.pokedexplus.poke_dex_plus_backend.model.EvolutionChain;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvolutionChainRepository extends JpaRepository<EvolutionChain, Long>{
+public interface EvolutionChainRepository extends JpaRepository<EvolutionChain, Integer>{
 }
